fix(style): 🐛 🎨 fix style bug
