+++ packages/web/src/pages/bounty/components/Payment/StageTerm.tsx
@@ -73,20 +73,20 @@ export default defineComponent({
           paymentInfo={this.item}
         />
 
-        <div class="border border-solid bg-[#F5F6FA] border-[#D5CFF4] rounded-sm mb-6 py-6 px-16 relative overflow-hidden">
+        <div class="border rounded-sm bg-[#F5F6FA] border-[#D5CFF4] mb-6 py-6 px-16 relative">
           {this.item?.status === 2 && <PayedMask />}
-          <span class="rounded-br-md bg-[#D5CFF4] py-1 px-2 top-0 left-0 text-[#3F2D99] absolute u-card-title2">
+          <span class="rounded-br-md bg-[#D5CFF4] text-primary py-1 px-2 top-[-1px] left-[-1px] absolute u-h5">
             {this.flagStr}
           </span>
           <div class="flex items-center">
-            <p class="flex-1 text-[#5331F4]">
-              <span class="text-[#5331F4] u-h3">{this.item.token1Amount || 0}</span>
-              <span class="mx-2 text-14px">{this.item.token1Symbol}</span>
+            <p class="flex-1 text-primary">
+              <span class="u-h2">{this.item.token1Amount || 0}</span>
+              <span class="ml-2 u-h6">{this.item.token1Symbol}</span>
               {this.item.token2Amount && (
                 <span>
-                  <span class="mx-2 text-20px">+</span>
-                  <span class="text-[#5331F4] u-h3">{this.item.token2Amount}</span>
-                  <span class="mx-2 text-14px">{this.item.token2Symbol}</span>
+                  <span class="font-sans mx-2 text-24px">+</span>
+                  <span class="u-h2">{this.item.token2Amount}</span>
+                  <span class="ml-2 u-h6">{this.item.token2Symbol}</span>
                 </span>
               )}
             </p>
@@ -97,7 +97,7 @@ export default defineComponent({
                 class={`h-9 w-30 font-primary font-semibold  -mr-8 px-8 ${
                   !this.payBtnAbled
                     ? 'text-[#FFFFFF]'
-                    : 'text-[#5331F4] border-1 border-[#5331F4] border-solid rounded-sm !hover:text-[#5331F4]'
+                    : 'text-primary border-1 border-[#5331F4] border-solid rounded-sm !hover:text-primary'
                 }`}
                 color={!this.payBtnAbled ? 'rgba(0,0,0,0.1)' : ''}
                 type="default"
