--- packages/web/src/pages/bounty/components/Payment/index.ts
@@ -1,1 +1,0 @@
-export { default as Payment } from './Panel'
